feat: creates ascii & git plugins
makes 2 plugins for the generator, ascii & git. there are a few ascii
generators but they don't play nice with the rest of the metalsmith
tools, so i decided to try my hand at my own. i might upstream some of
it if it makes sense .. because i have a different use case here than
what most probably have.

ascii

a standard ascii doc converter, takes .asciidoc, .adoc files and makes
them html. this one is a little different because it will also parse
metadata from the front. not sure the standard way to do this in
asciidoc but i think (and have used in the past) the `:key: value`
notation immediately at the top of the document, or right under the
first heading `=`. this plugin parses all the stuff before the first h2
heading `==` and sees if its a metadata tag.

git

just a simple plugin to test the waters, it removes all the git
repository files from the file tree so they don't get in the way of
anything.