chore: combining previous versions into one
this commit takes a couple of different site generators and combines
them into one generator. there is also something called `sxte` out there
that is fundamentally the same thing but in library form. i decided that
i do not want to make this thing a library (just yet) and instead have a
multi source project.

1. `src` is the main code for my site
2. `plugins` are specific things that do a specific function
3. `tools` are shared things that multiple plugins can use

ADDITIONS:

- adds the main site class with its required types and functions
- plugin(files): scans files, reads their content and adds them to the
  site
- plugin(frontmatter): reads the file and extracts frontmatter. expects the
  metadata to be at the beginning of the file
- plugin(templates): adds a templating engine that can be used for
  substitutions and generating new files
- plugin(markdown): adds a parser that converts markdown to html using
  the lib.haxe.org markdown library

ON PLUGIN OPTIONS:

the way i did most of these was to have a `typedef` and then an instance
that has all the defaults. every typedef was separated and all of them
needed to know what `area` they were working in, so that information was
duplicated.

the user could set the default area by changing a static inline
variable, but that felt bad, so now there is a function that will be
used to initalize the defaults, which can look at the main site settings
(which can be changed during initalization / runtime). so this feels
better.

i also wanted to use inheritance, which caused some issues, but now
there is a function that all defaults will have to use that will
initalize the values that are specific to the 'parent'.